rt: move `Runtime` into its own file
This removes the `Runtime` implementation from a cfg macro so it can be
formatted.
